explorando solucion 1
