+++ Makefile
@@ -8,7 +8,8 @@ SOURCES = main.c mooks.c parser/check_tree.c parser/constructors.c\
 parser/expansor_utils.c parser/expansor.c parser/free_tree.c parser/get_redir.c\
 parser/processline.c parser/str_utils.c executor/executor.c\
 executor/libexe.c executor/process.c non_pipable_builtin.c signal.c history.c\
-executor/apply_redirs.c GNL/get_next_line_utils.c GNL/get_next_line.c
+executor/apply_redirs.c GNL/get_next_line_utils.c GNL/get_next_line.c\
+parser/tmp_file_name.c
 OBJECTS = $(SOURCES:.c=.o)
 HEADERS = libft/headers/libft.h minishell.h executor.h
